+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * Fixed spectators will prevent arrows
 * Fixed editor value "spawn gold time"
 * Added option to disable separate chat
+* Added 1.16.x compatibility 
+* Added possibility to change table name of mysql database
 
 ### 1.5.5 Hotfix
 * Fixed language list getting